Move into packages so that the test package matches the main package structure
